[kern] add free_ept_page_lock and use it only out of spinlocks
free_ept_page may sleep because of set_page_dirty_lock and currently it
is called from atomic context, as can be proved by compiling the kernel
with CONFIG_DEBUG_ATOMIC_SLEEP=y. This can lead to soft lockups. Until
now, it hasn't happened but since introducing handling of SIGSTOP and
SIGCONT, the soft lockup happens when doing CTRL-Z and resume of a Dune
application.